+++ python/python/bystro/supervised_ppca/_misc_pt.py
@@ -66,7 +66,7 @@ def ldet_sw_full(A, U, B, V):
     ----------
     A : torch.Tensor
         Input matrix A.
-        
+
     U : torch.Tensor
         Input matrix U.
 
@@ -159,7 +159,7 @@ def inverse_sw_full(A, U, B, V):
 
 def ldet_sw_factor_analysis(Lambda, W, I_l=None):
     """
-    Compute the log determinant of the covariance matrix in a factor 
+    Compute the log determinant of the covariance matrix in a factor
     analysis model using the Sherman Woodbury identity
 
     Parameters
@@ -230,7 +230,7 @@ def inv_sw_factor_analysis(Lambda, W, I_l=None, I_p=None):
 
 def mvn_log_prob_sw(X, mu, Lambda, W, I_l=None):
     """
-    Compute the log probability of data points under a multivariate 
+    Compute the log probability of data points under a multivariate
     normal distribution using the Sherman Woodbury matrix identity.
 
     Parameters
@@ -248,13 +248,13 @@ def mvn_log_prob_sw(X, mu, Lambda, W, I_l=None):
         Weight matrix.
 
     I_l : torch.Tensor, optional
-        LxL identity matrix. If not provided, it is set to the 
+        LxL identity matrix. If not provided, it is set to the
         identity matrix.
 
     Returns
     -------
     torch.Tensor
-        Log probability of the data points under the multivariate 
+        Log probability of the data points under the multivariate
         normal distribution.
     """
     L, p = W.shape
@@ -280,7 +280,7 @@ def mvn_log_prob_sw(X, mu, Lambda, W, I_l=None):
     end = middle @ X_demeaned.T
 
     quad = -1 * X_demeaned.T * end
-    term3 = torch.sum(quad, axis=0) # type: ignore
+    term3 = torch.sum(quad, axis=0)  # type: ignore
 
     log_prob_window = term1 + term2 + term3 / 2
     log_prob = torch.mean(log_prob_window)
